+++ src/__tests__/integration/zoho/request.test.ts
@@ -16,12 +16,12 @@ let calendars: DAVCalendar[];
 
 beforeAll(async () => {
   authHeaders = getBasicAuthHeaders({
-    username: process.env.CREDENTIAL_BAIKAL_USERNAME,
-    password: process.env.CREDENTIAL_BAIKAL_PASSWORD,
+    username: process.env.CREDENTIAL_ZOHO_USERNAME,
+    password: process.env.CREDENTIAL_ZOHO_PASSWORD,
   });
   account = await createAccount({
     account: {
-      serverUrl: `${process.env.CREDENTIAL_BAIKAL_SERVER_URL}/dav.php`,
+      serverUrl: `${process.env.CREDENTIAL_ZOHO_SERVER_URL}`,
       accountType: 'caldav',
     },
     headers: authHeaders,
@@ -34,10 +34,13 @@ beforeAll(async () => {
 
 test('davRequest should be able to send normal webdav requests', async () => {
   const [result] = await davRequest({
-    url: `${process.env.CREDENTIAL_BAIKAL_SERVER_URL}/dav.php`,
+    url: process.env.CREDENTIAL_ZOHO_SERVER_URL,
     init: {
       method: 'PROPFIND',
-      headers: authHeaders,
+      headers: {
+        ...authHeaders,
+        depth: '0',
+      },
       namespace: 'd',
       body: {
         propfind: {
@@ -53,7 +56,7 @@ test('davRequest should be able to send normal webdav requests', async () => {
   expect(result.status).toBe(207);
   expect(result.statusText).toBe('Multi-Status');
   expect(result.ok).toBe(true);
-  expect(result.props?.currentUserPrincipal.href).toMatch(/dav.php\/principals\/.+/);
+  expect(result.props?.currentUserPrincipal.href).toMatch(/\/caldav\/.+\/user/);
   expect(Object.prototype.hasOwnProperty.call(result, 'raw')).toBe(true);
 });
 
@@ -65,10 +68,13 @@ test('davRequest should be able to send raw xml requests', async () => {
      </d:prop>
   </d:propfind>`;
   const [result] = await davRequest({
-    url: `${process.env.CREDENTIAL_BAIKAL_SERVER_URL}/dav.php`,
+    url: process.env.CREDENTIAL_ZOHO_SERVER_URL,
     init: {
       method: 'PROPFIND',
-      headers: authHeaders,
+      headers: {
+        ...authHeaders,
+        depth: '0',
+      },
       body: xml,
     },
     convertIncoming: false,
@@ -77,7 +83,7 @@ test('davRequest should be able to send raw xml requests', async () => {
   expect(result.status).toBe(207);
   expect(result.statusText).toBe('Multi-Status');
   expect(result.ok).toBe(true);
-  expect(result.props?.currentUserPrincipal.href).toMatch(/dav.php\/principals\/.+/);
+  expect(result.props?.currentUserPrincipal.href).toMatch(/\/caldav\/.+\/user/);
   expect(Object.prototype.hasOwnProperty.call(result, 'raw')).toBe(true);
 });
 
@@ -89,10 +95,13 @@ test('davRequest should be able to get raw xml response', async () => {
      </d:prop>
   </d:propfind>`;
   const [result] = await davRequest({
-    url: `${process.env.CREDENTIAL_BAIKAL_SERVER_URL}/dav.php`,
+    url: process.env.CREDENTIAL_ZOHO_SERVER_URL,
     init: {
       method: 'PROPFIND',
-      headers: authHeaders,
+      headers: {
+        ...authHeaders,
+        depth: '0',
+      },
       body: xml,
     },
     convertIncoming: false,
@@ -103,13 +112,13 @@ test('davRequest should be able to get raw xml response', async () => {
   expect(result.statusText).toBe('Multi-Status');
   expect(result.ok).toBe(true);
   expect(result.raw).toMatch(
-    /<d:current-user-principal><d:href>\/dav.php\/principals\/.+\/<\/d:href><\/d:current-user-principal>/
+    /<D:current-user-principal><D:href>\/caldav\/.+\/user\/<\/D:href><\/D:current-user-principal>/
   );
 });
 
 test('propfind should be able to find props', async () => {
   const [result] = await propfind({
-    url: `${process.env.CREDENTIAL_BAIKAL_SERVER_URL}/dav.php`,
+    url: process.env.CREDENTIAL_ZOHO_SERVER_URL,
     props: {
       [`${DAVNamespaceShort.DAV}:current-user-principal`]: {},
     },
@@ -120,14 +129,14 @@ test('propfind should be able to find props', async () => {
   expect(result.status).toBe(207);
   expect(result.statusText).toBe('Multi-Status');
   expect(result.ok).toBe(true);
-  expect(result.props?.currentUserPrincipal.href).toMatch(/dav.php\/principals\/.+/);
+  expect(result.props?.currentUserPrincipal.href).toMatch(/\/caldav\/.+\/user/);
   expect(Object.prototype.hasOwnProperty.call(result, 'raw')).toBe(true);
 });
 
 test('createObject should be able to create object', async () => {
   const iCalString = await fsp.readFile(`${__dirname}/../data/ical/9.ics`, 'utf-8');
 
-  const objectUrl = new URL('9.ics', calendars[0].url).href;
+  const objectUrl = new URL('b53b6846-ede3-4689-b744-aa33963e1586.ics', calendars[0].url).href;
   const response = await createObject({
     url: objectUrl,
     data: iCalString,
@@ -146,7 +155,9 @@ test('createObject should be able to create object', async () => {
   expect(response.ok).toBe(true);
   expect(calendarObject.url.length > 0).toBe(true);
   expect(calendarObject.etag.length > 0).toBe(true);
-  expect(calendarObject.data.split('\r\n').join('\n')).toEqual(iCalString);
+  expect(calendarObject.data.split('DTSTART:')[1].split('\n')[0]).toEqual('20210307T090800Z');
+  expect(calendarObject.data.split('DTEND:')[1].split('\n')[0]).toEqual('20210307T100800Z');
+  expect(calendarObject.data.split('SUMMARY:')[1].split('\n')[0]).toEqual('9');
 
   const deleteResult = await deleteObject({
     url: objectUrl,
@@ -160,7 +171,7 @@ test('updateObject should be able to update object', async () => {
   const iCalString = await fsp.readFile(`${__dirname}/../data/ical/10.ics`, 'utf-8');
   const updatedICalString = await fsp.readFile(`${__dirname}/../data/ical/11.ics`, 'utf-8');
 
-  const objectUrl = new URL('10.ics', calendars[0].url).href;
+  const objectUrl = new URL('f7fcf23d-a90a-4044-925f-91ad2c9c81cd.ics', calendars[0].url).href;
   const createResult = await createObject({
     url: objectUrl,
     data: iCalString,
@@ -171,16 +182,9 @@ test('updateObject should be able to update object', async () => {
   });
   expect(createResult.ok).toBe(true);
 
-  const [calendarObject] = await fetchCalendarObjects({
-    calendar: calendars[0],
-    objectUrls: [objectUrl],
-    headers: authHeaders,
-  });
-
   const updateResult = await updateObject({
     url: objectUrl,
     data: updatedICalString,
-    etag: calendarObject.etag,
     headers: {
       'content-type': 'text/calendar; charset=utf-8',
       ...authHeaders,
@@ -194,7 +198,11 @@ test('updateObject should be able to update object', async () => {
   });
 
   expect(result.ok).toBe(true);
-  expect((await result.text()).split('\r\n').join('\n')).toEqual(updatedICalString);
+
+  const iCalText = await result.text();
+
+  expect(iCalText.split('DTSTART:')[1].split('\n')[0]).toEqual('20210309T090800Z');
+  expect(iCalText.split('DTEND:')[1].split('\n')[0]).toEqual('20210309T100800Z');
 
   const deleteResult = await deleteObject({
     url: objectUrl,
@@ -207,7 +215,7 @@ test('updateObject should be able to update object', async () => {
 test('deleteObject should be able to delete object', async () => {
   const iCalString = await fsp.readFile(`${__dirname}/../data/ical/12.ics`, 'utf-8');
 
-  const objectUrl = new URL('test3.ics', calendars[0].url).href;
+  const objectUrl = new URL('0398667c-2292-4576-9751-a445f88394ab.ics', calendars[0].url).href;
   const createResult = await createObject({
     url: objectUrl,
     data: iCalString,